requirements.txt
PyYAML==6.0.1
requests==2.32.3
minio==7.2.7